refactor: `hf_utils` module
