+++ src/speaches/kokoro_utils.py
@@ -1,15 +1,18 @@
 from collections.abc import AsyncGenerator
 import logging
+from pathlib import Path
 import time
 from typing import Literal
 
+import httpx
+import huggingface_hub
 from kokoro_onnx import Kokoro
 import numpy as np
 
 from speaches.audio import resample_audio
+from speaches.hf_utils import list_model_files
 
-logger = logging.getLogger(__name__)
-
+KOKORO_REVISION = "c97b7bbc3e60f447383c79b2f94fee861ff156ac"
 SAMPLE_RATE = 24000  # the default sample rate for Kokoro
 Language = Literal["en-us", "en-gb", "fr-fr", "ja", "ko", "cmn"]
 LANGUAGES: list[Language] = ["en-us", "en-gb", "fr-fr", "ja", "ko", "cmn"]
@@ -27,6 +30,35 @@ VOICE_IDS = [
     "af_nicole",
     "af_sky",
 ]
+
+logger = logging.getLogger(__name__)
+
+
+def get_kokoro_model_path() -> Path:
+    file_name = "kokoro-v0_19.onnx"
+    onnx_files = list(list_model_files("hexgrad/Kokoro-82M", glob_pattern=f"**/{file_name}"))
+    if len(onnx_files) == 0:
+        raise ValueError(f"Could not find {file_name} file for 'hexgrad/Kokoro-82M' model")
+    return onnx_files[0]
+
+
+def download_kokoro_model() -> None:
+    model_id = "hexgrad/Kokoro-82M"
+    model_repo_path = Path(
+        huggingface_hub.snapshot_download(
+            model_id,
+            repo_type="model",
+            allow_patterns=["kokoro-v0_19.onnx"],
+            revision=KOKORO_REVISION,
+        )
+    )
+    # HACK
+    res = httpx.get(
+        "https://github.com/thewh1teagle/kokoro-onnx/releases/download/model-files/voices.bin", follow_redirects=True
+    ).raise_for_status()
+    voices_path = model_repo_path / "voices.bin"
+    voices_path.touch(exist_ok=True)
+    voices_path.write_bytes(res.content)
 
 
 async def generate_audio(
