+++ src/speaches/model_manager.py
@@ -13,7 +13,8 @@ from faster_whisper import WhisperModel
 from kokoro_onnx import Kokoro
 from onnxruntime import InferenceSession
 
-from speaches.hf_utils import get_kokoro_model_path, get_piper_voice_model_file
+from speaches.kokoro_utils import get_kokoro_model_path
+from speaches.piper_utils import get_piper_voice_model_file
 
 if TYPE_CHECKING:
     from collections.abc import Callable
