+++ tests/conftest.py
@@ -15,7 +15,7 @@ from pytest_mock import MockerFixture
 
 from speaches.config import Config, WhisperConfig
 from speaches.dependencies import get_config
-from speaches.hf_utils import download_kokoro_model
+from speaches.kokoro_utils import download_kokoro_model
 from speaches.main import create_app
 
 DISABLE_LOGGERS = ["multipart.multipart", "faster_whisper"]
